+++ java/src/main/java/com/codurance/training/tasks/AddTask.java
@@ -1,0 +1,5 @@
+package com.codurance.training.tasks;
+
+public interface AddTask {
+    public void add(String commandLine);
+}
