+++ java/src/main/java/com/codurance/training/tasks/CheckTask.java
@@ -1,0 +1,5 @@
+package com.codurance.training.tasks;
+
+public interface CheckTask {
+    public void check(String idString);
+}
