+++ java/src/main/java/com/codurance/training/tasks/DeleteTask.java
@@ -1,0 +1,5 @@
+package com.codurance.training.tasks;
+
+public interface DeleteTask {
+    public void delete(String stringId);
+}
