+++ java/src/main/java/com/codurance/training/tasks/ErrorInTask.java
@@ -1,0 +1,5 @@
+package com.codurance.training.tasks;
+
+public interface ErrorInTask {
+    public void error(String command);
+}
