+++ java/src/main/java/com/codurance/training/tasks/Help.java
@@ -1,0 +1,5 @@
+package com.codurance.training.tasks;
+
+public interface Help {
+    public void help();
+}
