+++ java/src/main/java/com/codurance/training/tasks/ShowTask.java
@@ -1,0 +1,5 @@
+package com.codurance.training.tasks;
+
+public interface ShowTask {
+    public void show();
+}
