+++ java/src/main/java/com/codurance/training/tasks/TaskSetDone.java
@@ -1,0 +1,5 @@
+package com.codurance.training.tasks;
+
+public interface TaskSetDone {
+    public void setDone(String idString, boolean done);
+}
