+++ java/src/main/java/com/codurance/training/tasks/UncheckTask.java
@@ -1,0 +1,5 @@
+package com.codurance.training.tasks;
+
+public interface UncheckTask {
+    public void uncheck(String idString);
+}
